server/requirements.txt
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
autoflake==2.2.0
black==23.7.0
click==8.1.6
exceptiongroup==1.1.3
fastapi==0.101.1
fastapi-pagination==0.12.8
greenlet==2.0.2
h11==0.14.0
idna==3.4
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
pyflakes==3.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==1.4.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
python-jose==3.3.0
pytest==7.2.1
pytest-cov==2.12.1
factory-boy==3.2.1
Faker==15.3.3
SQLAlchemy-Utils==0.37.9
httpx==0.24.1